The overwrite command was used to overwrite an existing command and create a new command. In this commit the functions are divided into two commands.
